requirements.txt
Flask
Flask-Login
Flask-WTF
WTForms
Werkzeug
Jinja2
python-dotenv
bcrypt
openai
langchain
langchain-openai
langchain-core
langgraph
PyPDF2
pdfplumber
python-docx
requests
beautifulsoup4
supabase
pinecone
google-generativeai
numpy
uvicorn
psycopg2-binary
pdfkit
flask_sqlalchemy
python-dateutil
xhtml2pdf
plasTeX